+++ src/orgs/smartsaver/sections/ufileresp/UfilerespLayout.js
@@ -1,0 +1,12 @@
+import React from 'react'
+import Layout from '../../../../components/Layout'
+
+const UfilerespLayout = ({ children }) => {
+  return (
+    <Layout>
+      <main>{children}</main>
+    </Layout>
+  )
+}
+
+export default UfilerespLayout
