+++ src/pages/smartsaver/ufileresp.js
@@ -1,7 +1,12 @@
 import React from 'react'
+import Layout from '../../orgs/smartsaver/sections/ufileresp/UfilerespLayout'
 
 const UfilerespPage = () => {
-  return <div className="UfilerespPage" />
+  return (
+    <Layout>
+      <h1 className="title">Ufileresp</h1>
+    </Layout>
+  )
 }
 
 export default UfilerespPage
